session_3/requirements.txt
asgiref==3.7.2
boto3==1.34.119
botocore==1.34.119
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-allauth==0.63.2
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.1.1
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
